pyproject.toml
[project]
name = "chat-with-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic==0.58.2",
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "python-multipart==0.0.20",
    "python-dotenv==1.1.1",
    "pytest>=8.4.1",
    "httpx>=0.24.0",
    "sqlalchemy>=2.0.43",
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "pyjwt>=2.10.1",
    "psycopg2-binary>=2.9.10",
    "stripe>=11.8.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-p", "no:warnings"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "api: API endpoint tests",
    "slow: Slow tests that may take longer to run"
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "datamodel-code-generator>=0.33.0",
    "deepeval>=3.6.0",
    "mypy>=1.17.1",
    "ruff>=0.12.9",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["B011"]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true
